requirements.txt
asgiref==3.8.1
Django==5.2.1
django-phonenumber-field==8.1.0
lxml==5.4.0
phonenumbers==9.0.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
